// backend/src/types/response/dedicatedPage/getDedicatedPageBannerResponse.ts
export type GetDedicatedPageBannerResponse = {
  id: number;
  general_banner_image: string | null;
  general_banner_fontsize: number | null;
  general_banner_font_color: string | null;
  general_banner_description_fontsize: number | null;
  general_banner_description_font_color: string | null;
  guest_logout_title_text: string | null;
  guest_logout_description_text: string | null;
  guest_register_button_text: string | null;
  guest_register_button_font_size: number | null;
  guest_register_button_font_color: string | null;
  guest_register_button_background_color: string | null;
  guest_login_button_text: string | null;
  guest_login_button_font_size: number | null;
  guest_login_button_font_color: string | null;
  guest_login_button_background_color: string | null;
  member_loggedIn_title_text: string | null;
  member_loggedIn_description_text: string | null;
  member_earn_button_text: string | null;
  member_earn_button_font_size: number | null;
  member_earn_button_font_color: string | null;
  member_earn_button_background_color: string | null;
  member_redeem_button_text: string | null;
  member_redeem_button_font_size: number | null;
  member_redeem_button_font_color: string | null;
  member_redeem_button_background_color: string | null;
  status: string | null;
  user_id: number;
  admin_ref: number | null;
  created_by: number | null;
  created_at: Date | null;
  updated_at: Date | null;
  updated_by: number | null;
};