// backend/src/types/response/dedicatedPage/getDedicatedPageExplainerResponse.ts
export type GetDedicatedPageExplainerResponse = {
  id: number;
  general_background_color: string | null;
  general_header_title_font_size: number | null;
  general_header_title_font_color: string | null;
  general_header_subtitle_font_size: number | null;
  general_header_subtitle_font_color: string | null;
  guest_header_text: string | null;
  guest_step_number_font_size: number | null;
  guest_step_number_font_color: string | null;
  guest_step_title_font_size: number | null;
  guest_step_title_font_color: string | null;
  guest_step_subtitle_font_size: number | null;
  guest_step_subtitle_font_color: string | null;
  guest_step_1_title_text: string | null;
  guest_step_1_subtitle_text: string | null;
  guest_step_2_title_text: string | null;
  guest_step_2_subtitle_text: string | null;
  guest_step_3_title_text: string | null;
  guest_step_3_subtitle_text: string | null;
  member_header_text: string | null;
  member_subtitle_text: string | null;
  member_activity_action_text: string | null;
  member_points_text: string | null;
  member_data_text: string | null;
  member_your_reward_title_text: string | null;
  member_reward_title_font_size: number | null;
  member_reward_title_font_color: string | null;
  member_no_reward_subtitle_text: string | null;
  member_has_reward_subtitle_text: string | null;
  member_reward_subtitle_font_size: string | null;
  member_reward_subtitle_font_color: string | null;
  status: string | null;
  user_id: number;
  admin_ref: number | null;
  created_by: number | null;
  created_at: Date | null;
  updated_at: Date | null;
  updated_by: number | null;
};