// backend/src/types/response/dedicatedPage/getDedicatedPageWayToEarnResponse.ts
export type GetDedicatedPageWayToEarnResponse = {
  id: number;
  general_background_color: string | null;
  general_title_text: string | null;
  general_subtitle_text: string | null;
  general_title_font_size: number | null;
  general_title_font_color: string | null;
  general_subtitle_font_size: number | null;
  general_subtitle_font_color: string | null;
  status: string | null;
  user_id: number;
  admin_ref: number | null;
  created_by: number | null;
  created_at: Date | null;
  updated_at: Date | null;
  updated_by: number | null;
};